+++ src/Engine/View/ConsoleViewStrategy.h
@@ -18,7 +18,7 @@ namespace Engine
             void PreRender() override;
             void Render(const Shape& shape) override;
             void PostRender() override;
-            sf::Vector2i GetMousePosition() override { return { 0,0 }; }; // placeholder
+            sf::Vector2f GetMousePosition() override { return { 0,0 }; }; // placeholder
         };
     };
 };
